Refactor: Audit fixes, Performance improvements, and Type safety
