+++ .dockerignore
@@ -1,0 +1,10 @@
+node_modules
+dist
+.git
+.github
+.DS_Store
+.env
+.env.*
+*.log
+coverage
+tests
