+++ src/components/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '@/contexts/AuthContextNew';
+import { useAuth } from "@/contexts/AuthContext";
 import { Loader2 } from 'lucide-react';
 
 interface ProtectedRouteProps {
