+++ src/components/SEO.tsx
@@ -1,0 +1,42 @@
+import { Helmet } from 'react-helmet-async';
+
+interface SEOProps {
+    title: string;
+    description: string;
+    canonical?: string;
+    type?: 'website' | 'article';
+    image?: string;
+}
+
+export const SEO = ({
+    title,
+    description,
+    canonical,
+    type = 'website',
+    image = '/og-image.png'
+}: SEOProps) => {
+    const siteTitle = 'Keen VPN';
+    const fullTitle = `${title} | ${siteTitle}`;
+
+    return (
+        <Helmet>
+            {/* Standard metadata */}
+            <title>{fullTitle}</title>
+            <meta name='description' content={description} />
+            {canonical && <link rel="canonical" href={canonical} />}
+
+            {/* Open Graph / Facebook */}
+            <meta property='og:type' content={type} />
+            <meta property='og:title' content={fullTitle} />
+            <meta property='og:description' content={description} />
+            <meta property='og:image' content={image} />
+
+            {/* Twitter */}
+            <meta name='twitter:card' content='summary_large_image' />
+            <meta name='twitter:creator' content='@keenvpn' />
+            <meta name='twitter:title' content={fullTitle} />
+            <meta name='twitter:description' content={description} />
+            <meta name='twitter:image' content={image} />
+        </Helmet>
+    );
+};
