+++ vite.config.ts
@@ -1,9 +1,11 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
 import path from "path";
+import viteCompression from 'vite-plugin-compression';
+import Sitemap from 'vite-plugin-sitemap';
 
 // https://vitejs.dev/config/
-export default defineConfig(({ mode }) => ({
+export default defineConfig({
   server: {
     host: "::",
     port: 8080,
@@ -11,13 +13,12 @@ export default defineConfig(({ mode }) => ({
   base: "./",
   plugins: [
     react(),
+    viteCompression(),
+    Sitemap({ hostname: 'https://vpnkeen.com' }),
   ],
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
     },
   },
-  define: {
-    __WS_TOKEN__: JSON.stringify(process.env.WS_TOKEN || ''),
-  },
-}));
+});
